+++ apps/backend/src/modules/ai-assistant/libs/helpers/helpers.ts
@@ -1,0 +1,16 @@
+export {
+	generateChangeTaskSuggestionsResponse,
+	runChangeTaskByCategoryOptions,
+} from "./change-task/run-change-task-by-category-options.helper.js";
+export {
+	generateExplainTaskSuggestionsResponse,
+	runExplainTaskOptions,
+} from "./explain-task/run-explain-task-option.helper.js";
+export {
+	generateQuestionsAnswersPrompt,
+	generateUserScoresPrompt,
+} from "./initial-chat/initial-chat.js";
+export {
+	generateTaskSuggestionsResponse,
+	runSuggestTaskByCategoryOptions,
+} from "./suggest-task-by-category/run-suggest-task-by-category-option.helper.js";