+++ apps/frontend/src/modules/quiz/slices/quiz.slice.ts
@@ -52,7 +52,7 @@ const { actions, name, reducer } = createSlice({
 			state.dataStatus = DataStatus.FULFILLED;
 			state.questionsByCategories = action.payload.items;
 			state.currentCategoryQuestions =
-				state.questionsByCategories[state.currentCategoryIndex] || null;
+				state.questionsByCategories[state.currentCategoryIndex] ?? null;
 		});
 		builder.addCase(getAllQuestions.rejected, (state) => {
 			state.dataStatus = DataStatus.REJECTED;
@@ -66,7 +66,7 @@ const { actions, name, reducer } = createSlice({
 			state.dataStatus = DataStatus.FULFILLED;
 			state.questionsByCategories = action.payload.items;
 			state.currentCategoryQuestions =
-				state.questionsByCategories[state.currentCategoryIndex] || null;
+				state.questionsByCategories[state.currentCategoryIndex] ?? null;
 		});
 		builder.addCase(getQuestionsByCategoryIds.rejected, (state) => {
 			state.dataStatus = DataStatus.REJECTED;
@@ -127,6 +127,9 @@ const { actions, name, reducer } = createSlice({
 	initialState,
 	name: "quiz",
 	reducers: {
+		editScore(state, action: PayloadAction<QuizScoresGetAllItemResponseDto[]>) {
+			state.scores = action.payload;
+		},
 		nextQuestion(state) {
 			state.currentCategoryIndex += PREVIOUS_INDEX_OFFSET;