+++ packages/shared/src/modules/tasks/libs/enums/tasks-api-path.enum.ts
@@ -5,6 +5,7 @@ const TasksApiPath = {
 	NOTES: "/notes",
 	NOTES_$ID: "/notes/:id",
 	PAST: "/past",
+	ROOT: "/",
 } as const;
 
 export { TasksApiPath };