+++ packages/shared/src/modules/tasks/libs/types/task-create-dto.type.ts
@@ -1,0 +1,8 @@
+type TaskCreateDto = {
+	categoryId: number;
+	categoryName: string;
+	description: string;
+	dueDate: string;
+	label: string;
+};
+export { type TaskCreateDto };