+++ packages/shared/src/modules/tasks/tasks.ts
@@ -1,6 +1,8 @@
 export { TasksApiPath, TaskStatus } from "./libs/enums/enums.js";
 export {
+	type TaskCreateDto,
 	type TaskDto,
+	type TaskGetAllResponseDto,
 	type TaskNoteDto,
 	type TaskNoteParametersDto,
 	type TaskNoteRequestDto,